fix undefined behavior

+++ src/Functions/bitSlice.cpp
@@ -5,7 +5,6 @@
 #include <Functions/FunctionFactory.h>
 #include <Functions/FunctionHelpers.h>
 #include <Functions/GatherUtils/Algorithms.h>
-#include <Functions/GatherUtils/GatherUtils.h>
 #include <Functions/GatherUtils/Sinks.h>
 #include <Functions/GatherUtils/Slices.h>
 #include <Functions/GatherUtils/Sources.h>
@@ -132,7 +131,7 @@ public:
                         source, StringSink(*col_res, input_rows_count), static_cast<size_t>(start_value - 1));
                 else if (start_value < 0)
                     bitSliceFromRightConstantOffsetUnbounded(
-                        source, StringSink(*col_res, input_rows_count), static_cast<size_t>(-start_value));
+                        source, StringSink(*col_res, input_rows_count), -static_cast<size_t>(start_value));
                 else
                     throw Exception("Indices in strings are 1-based", ErrorCodes::ZERO_ARRAY_OR_TUPLE_INDEX);
             }
@@ -150,7 +149,7 @@ public:
                         source, StringSink(*col_res, input_rows_count), static_cast<size_t>(start_value - 1), length_value);
                 else if (start_value < 0)
                     bitSliceFromRightConstantOffsetBounded(
-                        source, StringSink(*col_res, input_rows_count), static_cast<size_t>(-start_value), length_value);
+                        source, StringSink(*col_res, input_rows_count), -static_cast<size_t>(start_value), length_value);
                 else
                     throw Exception("Indices in strings are 1-based", ErrorCodes::ZERO_ARRAY_OR_TUPLE_INDEX);
             }
@@ -365,7 +364,7 @@ public:
             if (start && length)
             {
                 bool left_offset = start > 0;
-                size_t offset = left_offset ? start - 1 : -start;
+                size_t offset = left_offset ? static_cast<size_t>(start - 1) : -static_cast<size_t>(start);
                 size_t size = src.getElementSize();
 
                 size_t offset_byte;
